ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
imagehash==4.3.1
requests==2.32.3
